==> 02_prochazeni_dat_web_scraping.ipynb <==
Procházení dat v základně pročištěném datasetu

import pandas as pd
import json
import collections

cafes = pd.read_json('dataset_google-maps-extractor_Brno_kavarny_cleaned.json')
cafes.head()

categories = cafes.groupby('categoryName')['title'].count().to_dict()

with open('list_of_categories.json', mode='w', encoding='utf-8') as output_file:
    json.dump(categories, output_file, indent=4, ensure_ascii=False)

cafes.columns

cafes.shape

list_of_postalCodes = cafes['postalCode']
list_of_unique_postalCodes = cafes['postalCode'].unique()
print(list_of_unique_postalCodes)

list_of_postalCodes.info

postalCode_counts = collections.Counter(list_of_postalCodes)
postalCode_counts

category_cafe = pd.read_json('dataset_google-maps-extractor_Brno_category_cafe_cleaned.json')
category_cafe.shape
# category_cafe

category_cafe['categoryName'].unique()

not_cafe = category_cafe[category_cafe['categoryName'] != 'Cafe']
not_cafe = not_cafe[category_cafe['categoryName'] != 'Coffee shop'] 
not_cafe = not_cafe[category_cafe['categoryName'] != 'Espresso bar']
not_cafe

category_coffee_shop = pd.read_json('dataset_google-maps-extractor_Brno_category_coffee_shop_cleaned.json')
category_coffee_shop.shape
category_coffee_shop.head()

category_coffee_shop['categoryName'].unique()

not_cafe_2 = category_coffee_shop[category_coffee_shop['categoryName'] == 'Wholesale bakery']
not_cafe_2

category_espresso_bar = pd.read_json('dataset_google-maps-extractor_Brno_category_espresso_bar_cleaned.json')
category_espresso_bar.shape
category_espresso_bar.head()

category_espresso_bar['categoryName'].unique()

not_cafe_3 = category_espresso_bar[category_espresso_bar['categoryName'] != 'Espresso bar']
not_cafe_3 = not_cafe_3[category_espresso_bar['categoryName'] != 'Cafe']
not_cafe_3 = not_cafe_3[category_espresso_bar['categoryName'] != 'Coffee shop']
not_cafe_3
