==> 03_final_dataset.ipynb <==
import pandas as pd
import json
import collections

Závěrečné propojení datasetů:

with open('dataset_google-maps-extractor_Brno_kavarny_cleaned.json', encoding= 'utf-8') as input_file_1:
    cafes_main = json.load(input_file_1)

with open('diff_cafe.json', encoding= 'utf-8') as input_file_2:
    cafes_2 = json.load(input_file_2)

with open('diff_coffee_shop.json', encoding= 'utf-8') as input_file_3:
    cafes_3 = json.load(input_file_3)

with open('diff_espresso_bar.json', encoding= 'utf-8') as input_file_4:
    cafes_4 = json.load(input_file_4)


cafes_count = len(cafes_main) + len(cafes_2) + len(cafes_3) + len(cafes_4) 
cafes_count

final_dataset = cafes_main + cafes_2 + cafes_3 + cafes_4

total_cafes_count = len(final_dataset)
total_cafes_count

with open('maps_google_scraped_final_dataset_V2.json', mode= 'w', encoding= 'utf-8') as output_file:
    json.dump(final_dataset, output_file, indent= 4, ensure_ascii= False)

Procházení a kontrola propojeného datasetu:

final_read = pd.read_json('maps_google_scraped_final_dataset_V2.json')
final_read.info()
final_read.head(5)

unique_titles = final_read['title'].unique()

nbr_titles = len(unique_titles)
nbr_titles


titles = dict()

for cafe in final_dataset:
    if cafe['title'] not in titles.keys():
        titles[cafe['title']] = 1
    else:
        titles[cafe['title']] += 1

titles

for title, number in titles.items():
    if number > 1:
        print(title, number)

for cafe in final_dataset:
    if cafe['title'] == 'Bukovský kavárna a cukrářství':
        print(cafe)


for cafe in final_dataset:
    if cafe['title'] == 'Trdlokafe':
        print(cafe)

for cafe in final_dataset:
    if cafe['title'] == 'KOFI-KOFI':
        print(cafe)

for cafe in final_dataset:
    if cafe['title'] == 'Starbucks':
        print(cafe)

for cafe in final_dataset:
    if cafe['title'] == 'Costa Coffee':
        print(cafe)

list_of_placeID = final_read['placeId']
list_of_placeID.info()
unique = list_of_placeID.unique()
len(unique)

placeID = collections.Counter(list_of_placeID)

not_unique = []

for id, count in placeID.items():
    if count > 1:
        not_unique.append(id)

print(not_unique)

null_postalCode = final_read[final_read['postalCode'].isna()]
null_postalCode
