==> 06_filtrace_podniku.ipynb <==
Script pro další pročištění dat

*I po základním pročištění v datech zůstalo spoustu podniků, které nejsou kavárnami. Podle námi stanoveného kritéria pro určení kavárny a podle dalších klíčů jsme sestavily seznam non_cafes_final.txt, tento načetly do Pythonu a podniky z dat vymazaly.*

import pandas as pd

data = pd.read_csv('maps_google_scraped_final_dataset_V5.csv')
data.columns


data = data.drop('Unnamed: 0', axis= 1)
data.head(5)

with open('non_cafes_final.txt', mode='r', encoding='utf-8') as file:
    non_cafes = file.readlines()

non_cafes


non_cafes_cleaned = []

for cafe in non_cafes:
    cafe = cafe.strip()
    non_cafes_cleaned.append(cafe)

non_cafes_cleaned

condition = (data['title'].isin(non_cafes_cleaned))

data_filtered = data[~condition]

data_filtered = data_filtered.set_index('placeId')

data_filtered.shape

data_filtered.to_csv('maps_google_scraped_final_dataset_V7.csv')
