==> 07_souradnice.ipynb <==
Script pro výpočet vzdáleností mezi vysokoškolskými fakultami a kavárnami (podle souřadnic)

import pandas as pd
from geopy.distance import distance
import json

Definice funkce pro výpočet vzdálenosti:

def calculate_distance(row):
    coords_1 = (row['lat_x'], row['lng_x'])
    coords_2 = (row['lat_y'], row['lng_y'])
    return distance(coords_1, coords_2).meters

data_cafes = pd.read_csv('maps_google_scraped_final_dataset_V6.csv')
data_cafes.columns


print(data_cafes['location'].apply(type).unique())

loc_data_cafes = data_cafes['location'].apply(lambda x: json.loads(x.replace("'", '"')))

locations = pd.json_normalize(loc_data_cafes)
locations

loc_data_cafes_normalized = pd.concat([data_cafes.drop('location', axis=1), locations], axis=1)
loc_data_cafes_normalized = loc_data_cafes_normalized[['placeId', 'title', 'lat', 'lng']]
loc_data_cafes_normalized

loc_data_cafes_normalized.to_csv('location_cafes.csv')

data_faculties = pd.read_csv('Fakulty_Brno.csv')
data_faculties.columns

data_faculties.rename(columns={'X': 'lng', 'Y': 'lat'}, inplace=True)
loc_data_faculties = data_faculties[['objectid', 'name', 'lat', 'lng']]
loc_data_faculties

CROSS JOIN (spojí každý řádek jednoho DataFrame s každým řádkem druhého DataFrame a vytvoří tak všechny možné kombinace řádků mezi oběma datovými sadami):

cross_joined_loc_data = pd.merge(loc_data_cafes_normalized, loc_data_faculties, how='cross')
cross_joined_loc_data

Výpočet vzdálenosti pomocí vlastní funkce calculate_distance():

cross_joined_loc_data['distance_m'] = cross_joined_loc_data.apply(calculate_distance, axis=1)
cross_joined_loc_data['distance_m']

Uložení výsledků do souboru:

print(cross_joined_loc_data[['placeId', 'title', 'objectid', 'name', 'distance_m']])

cross_joined_loc_data.to_csv('distance_cafe_faculty.csv')
